Separate Dom from logic
