+++ .eslintrc.json
@@ -14,7 +14,8 @@
       "no-shadow": "off",
       "no-param-reassign": "off",
       "eol-last": "off",
-      "arrow-parens": "off"
+      "arrow-parens": "off",
+      "import/extensions": 0
     },
     "ignorePatterns": [
       "dist/",
