+++ index.html
@@ -64,8 +64,7 @@
   <script src="https://www.gstatic.com/firebasejs/8.2.4/firebase-analytics.js"></script>
   <script src="https://www.gstatic.com/firebasejs/8.2.4/firebase-database.js"></script>
 
-  <script src="js/library.js"></script>
-
+    <script type="module" src="js/library.js"></script>
 
 </body>
