+++ js/firebase.js
@@ -1,0 +1,12 @@
+const firebaseConfig = {
+  apiKey: 'AIzaSyAkEY49W_RKvLI6rcO_r6SjnrJUvwduu2U',
+  authDomain: 'library-836fa.firebaseapp.com',
+  databaseURL: 'https://library-836fa-default-rtdb.firebaseio.com',
+  projectId: 'library-836fa',
+  storageBucket: 'library-836fa.appspot.com',
+  messagingSenderId: '742428988728',
+  appId: '1:742428988728:web:9081a66a88d2169030ddcd',
+  measurementId: 'G-7T30W83E97',
+};
+
+export default firebaseConfig;
